Disable long_to_short_function_def in macros/exprs
