+++ test/runtests.jl
@@ -19,6 +19,7 @@ Pkg.develop(; path = dirname(@__DIR__))
     include("options/annotate_untyped_fields_with_any.jl")
     include("options/for_to_in.jl")
     include("options/import_to_using.jl")
+    include("options/long_to_short_function_def.jl")
     include("options/short_circuit_to_if.jl")
     include("options/trailing_comma.jl")
     include("options/v2_stable_multiline_strings.jl")
